pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Daiki.Ichihara <icchi314@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
python-dotenv = "^1.0.0"
azure-cognitiveservices-speech = "^1.28.0"
openai = "^0.27.7"
ruff = "^0.0.270"
flask-cors = "^3.0.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = [
    # pyflakes
    "F",
    # pycodestyle errors, warnings
    "E",
    "W",
    # isort
    # "I",
    # flake8-2020
    "YTT",
    # flake8-bugbear
    "B",
]
 
target-version = "py311"
 
line-length = 120